movre file to admin folder and update links
