+++ client/src/Components/Layout/Base/BaseLayout.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import { Grid } from '@material-ui/core';
-import AddNewEvent from './../AddNewEvent';
+import AddNewEvent from '../admin/AddNewEvent';
 import Home from '../Home';
 import Header from '../../Common/Header/Header';
 import SignUp from '../SignUp/SignUp';
@@ -9,7 +9,7 @@ import EventDetails from '../EventDetails';
 import Login from '../Login/Login';
 import Profile from '../Profile/Profile';
 import TakeAttendance from '../../Layout/admin/Attendance/TakeAttendance';
-import AdminEventDetails from '../AdminEventDetails';
+import AdminEventDetails from '../admin/AdminEventDetails';
 import AdminHome from '../../Layout/admin/Home/Home';
 import AuthRouter from './AuthRouter';
 import isAuth from '../../../helpers/isAuth';
